Remove unused imports and fixtures,
update test methods with new class instantiation.
