+++ tests/test_processors_ayon_hiero_effect_file.py
@@ -1,6 +1,7 @@
 import pytest
 from pathlib import Path
 
+from lablib.processors import AYONHieroEffectsFileProcessor
 
 from tests.lib.testing_classes import MainTestClass
 
@@ -14,7 +15,13 @@ class TestAYONHieroEffectFileProcessor(MainTestClass):
             "BLD_010_0010_effectPlateMain_v000.json"
         ]
     )
-    def test_EffectsFileProcessor(self, path, effect_processor):  # noqa: F811
+    def test_EffectsFileProcessor(self, path):  # noqa: F811
+
+        effect_processor = AYONHieroEffectsFileProcessor(
+            Path(path)
+        )
+        self.log.debug(f"{effect_processor = }")
+        effect_processor.load()
 
         self.log.debug(effect_processor.color_operators)
         assert effect_processor.filepath == Path(path)
